use `long long` instead of `int`

+++ T7.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 void T7() {
     int a, n;
-    int ans = 0, temp;
+    long long ans = 0, temp;
     cout << "Enter two numbers, it will run like a+aa+aaa+..." << endl;
     cin >> a >> n;
     for (int i = 0; i <= n; i++) {
